Ajout des collisions / Événements de jeu / Spawn des astéroïdes
- Ajout des événements de jeu via GameEvents
- Ajout des détections de collisions, utilisées via @game.on_collide
- Ajout du spawn des astéroïdes

A faire :
- Profiling (temps d'exécution de chaque tronçon de code)
- Unit Tests via d'autres fichiers
- Tests de distribution / environnement
- (déjà commencé dans le dossier collisions)  Améliorer l'efficacité des détections de collisions en utilisant des quadTrees
 => Implémenter les différentes classes de BoundingBox
- Améliorer l'efficacité de l'affichage
  => En n'affichant que les entités à l'écran (utilisation des BoundingBox)
  => En utilisant des batchs ? PLS
- Rajouter des commentaires / explications

+++ entity.py
@@ -41,8 +41,8 @@ class Entity:
 
     def __init__(self, pos, game_state, speed = np.array([0, 0])):
 
-        self._pos = np.array(pos)
-        self.speed = np.array(speed)
+        self._pos = np.array(pos).astype(float)
+        self.speed = np.array(speed).astype(float)
 
         self._game_state = game_state
         self.camera = game_state.camera
